Added database support - postgres && sqlite3 as backup
Added transactional insert batching

Adds intial engine/gRPC functionality

Increase deadline timeout to 1 minute

Add OHLC grpc command with POC code

add logic to not overwrite current configtest.json

Add comment due to CI gosec linter alert

+++ exchanges/bittrex/bittrex_types.go
@@ -94,7 +94,7 @@ type MarketHistory struct {
 	Success bool   `json:"success"`
 	Message string `json:"message"`
 	Result  []struct {
-		ID        int     `json:"Id"`
+		ID        int64   `json:"Id"`
 		Timestamp string  `json:"TimeStamp"`
 		Quantity  float64 `json:"Quantity"`
 		Price     float64 `json:"Price"`
